automatically backed up by learn

+++ db/migrate/user.rb
@@ -4,7 +4,7 @@ class CreateUserTable < ActiveRecord::Migration[4.2]
     create_table :users do |t|
       t.string :username
       t.string :password
-      t.float :balance
+      t.decimal :balance
     end
   end
 end
